% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download CSV data}
\usage{
download_data(exportname, data)
}
\arguments{
\item{exportname}{name of the CSV file.}

\item{data}{data to be download.}
}
\description{
This function is designed to download CSV data
}
