% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_plot.R
\name{download_plot}
\alias{download_plot}
\title{Download static plot}
\usage{
download_plot(exportname, plot)
}
\arguments{
\item{exportname}{name of the CSV file.}

\item{plot}{plot to be download.}
}
\description{
This function is designed to download plots
}
