% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_s3_parquet_data.R
\name{download_s3_parquet_data}
\alias{download_s3_parquet_data}
\title{Download S3 parquet data}
\usage{
download_s3_parquet_data(
  bucket_name,
  prefix_name,
  aws_access_key_id,
  aws_secret_access_key,
  partitions
)
}
\arguments{
\item{bucket_name}{name of the s3 bucket.}

\item{prefix_name}{prefix used on the s3 bucket.}

\item{aws_access_key_id}{AWS S3 bucket secret key id.}

\item{aws_secret_access_key}{AWS S3 bucket secret access key.}

\item{partitions}{number of parquet partitions within s3 bucket.}
}
\description{
This function will download parquet data from S3 bucket and it will convert it to tibble for further analytics.
}
