% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stop_words.R
\name{generate_stop_words}
\alias{generate_stop_words}
\title{Stop words list}
\usage{
generate_stop_words(list, type_list = "site")
}
\arguments{
\item{list}{list of stop words elements.}

\item{type_list}{type of analysis. By site or by language.}
}
\description{
This function process free text data into a corpus for further analytics.
}
