% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_processed_corpus.R
\name{get_processed_corpus}
\alias{get_processed_corpus}
\title{Process corpus}
\usage{
get_processed_corpus(
  id_field = NULL,
  column_name,
  dataset,
  ngrams = 2,
  stop_word_lang,
  dictionary_vector
)
}
\arguments{
\item{id_field}{column to be used as document id.}

\item{column_name}{name of the s3 bucket.}

\item{dataset}{prefix used on the s3 bucket.}

\item{ngrams}{number of grams to be used for analysis.}

\item{stop_word_lang}{vector containing stop words.}

\item{dictionary_vector}{vector containing data dictionary to be removed.}
}
\description{
This function process free text data into a corpus for further analytics.
}
