% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_to_col.R
\name{index_to_col}
\alias{index_to_col}
\title{Create a column index}
\usage{
index_to_col(data, column_name)
}
\arguments{
\item{data}{data to be processed.}

\item{column_name}{index column name.}
}
\description{
This function will create a column as index for further analytics.
}
