% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_s3_parquet_update.R
\name{latest_s3_parquet_update}
\alias{latest_s3_parquet_update}
\title{Get latest S3 data update}
\usage{
latest_s3_parquet_update(
  bucket_name,
  prefix_name,
  aws_access_key_id,
  aws_secret_access_key,
  time_zone = "America/Los_Angeles"
)
}
\arguments{
\item{bucket_name}{name of the s3 bucket.}

\item{prefix_name}{prefix used on the s3 bucket.}

\item{aws_access_key_id}{AWS S3 bucket secret key id.}

\item{aws_secret_access_key}{AWS S3 bucket secret access key.}

\item{time_zone}{time zone to convert for date update (default Los Angeles Pacific Time).}
}
\description{
This function will provide the latest date when S3 bucket has been updated.
}
