% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_main_box.R
\name{lp_main_box}
\alias{lp_main_box}
\title{Create main box}
\usage{
lp_main_box(title_box, image_name, button_name, description)
}
\arguments{
\item{title_box}{title of the box.}

\item{image_name}{name of the image to be added to the box.}

\item{button_name}{name of the box button.}

\item{description}{brief description of the box.}
}
\description{
This function creates a main box for UI.
}
