% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nodata.R
\name{plot_nodata}
\alias{plot_nodata}
\title{Plot no data available}
\usage{
plot_nodata(height_plot = 450, message = "No data available", text_size = 20)
}
\arguments{
\item{height_plot}{value showing the height of the plot.}

\item{message}{message to be used by the function when data is not available.}

\item{text_size}{text size for the message.}
}
\description{
This function will show a message when there's no data to populate a plot
}
