% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Visualize a ROC plot}
\usage{
plot_roc(
  tag,
  score,
  model_name = NA,
  subtitle = NA,
  interval = 0.2,
  plotly = FALSE
)
}
\arguments{
\item{tag}{the predicted value.}

\item{score}{the actual value.}

\item{model_name}{the name of the model to be added in the plot.}

\item{subtitle}{the subtitle of the model to added in the plot.}

\item{interval}{interval used in the axis.}

\item{plotly}{if want to convert plot into plotly.}
}
\description{
This function is designed to display a ROC plot using Takeda's colors.
}
