% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roldanpack-package.R
\docType{package}
\name{roldanpack-package}
\alias{roldanpack}
\alias{roldanpack-package}
\title{roldanpack: Personal R Packages of Alexis Roldan}
\description{
Here you will find most R functions that I use for my personal use.
}
\author{
\strong{Maintainer}: Alexis Roldan \email{alexis.m.roldan.ds@gmail.com}

}
\keyword{internal}
