% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savechart_button.R
\name{savechart_button}
\alias{savechart_button}
\title{Save chart button}
\usage{
savechart_button(
  outputId,
  label = "Save chart",
  class = NULL,
  disabled = FALSE,
  icon_image = "image"
)
}
\arguments{
\item{outputId}{output id for the button.}

\item{label}{label for the button.}

\item{class}{button class type.}

\item{disabled}{boolean if button is disable or enabled.}

\item{icon_image}{name of the icon image to be displayed.}
}
\description{
This function will create a CSS script for save chart button
}
