% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_highlight.R
\name{search_highlight}
\alias{search_highlight}
\title{Create a highlight search function}
\usage{
search_highlight(column, pattern)
}
\arguments{
\item{column}{column to highlight.}

\item{pattern}{pattern to highlight within column.}
}
\description{
This function will highlight query within specific column.
}
