% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words.R
\name{search_words}
\alias{search_words}
\title{Create a search word engine}
\usage{
search_words(data, token_list, query, id, list_n = 3, stop_words, n_top = 200)
}
\arguments{
\item{data}{data to be processed.}

\item{token_list}{list of processed tokens.}

\item{query}{keywords to search.}

\item{id}{id column from dataset.}

\item{list_n}{number of dataset within list}

\item{stop_words}{list of stopwords (vector).}

\item{n_top}{max number of results.}
}
\description{
This function look for specific keywords within dataset.
}
