% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordHighlight.R
\name{wordHighlight}
\alias{wordHighlight}
\title{Highlight word in table}
\usage{
wordHighlight(word_highlight, color_highlight = "yellow")
}
\arguments{
\item{word_highlight}{word to be highlighted.}

\item{color_highlight}{color used to highlight the word.}
}
\description{
This function is to create a CSS script to highlight a specific word
}
