% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box.R
\name{value_box}
\alias{value_box}
\title{Download static plot}
\usage{
value_box(value, subtitle, icon, color, width = 3)
}
\arguments{
\item{value}{value to be added to the box.}

\item{subtitle}{subtitle for the box.}

\item{icon}{select the icon for the box}

\item{color}{choose the color of the box.}

\item{width}{width size for the box.}
}
\description{
This function is designed to download plots
}
